alternance-cours/semaine2/requirements.txt
streamlit
mysql-connector-python
geocoder
geopy
requests